Added sunsep and moonsep convenience functions
Added convenience functions for printing daily Sun and Moon separations every N days.
Added function to plot Sun separations over the course of a year, with a minimum separation limit line.
 Updated example docs to include these.
Made some tweaks to the readthedocs files -- moved the .rst example files to an archive, and instead soft-linked the .md versions in the examples folder.  This should help streamline things when making updates to docs.